version2/requirements.txt
faicons
shiny
seaborn
opencv-python==4.8.1.78
pyautogui
seaborn
matplotlib
pandas
plotly.express
